+++ python/tests/web_editor.py
@@ -22,3 +22,11 @@ def test_change_username(app_admin):
     assert rv.status_code == 200
     rv = app_admin.get('/auth/account')
     assert b'admin-tmp' not in rv.data
+
+def test_username_redirect(app_admin):
+
+    rv = app_admin.get('/u/admin')
+    assert rv.status_code == 302
+
+    rv = app_admin.get('/u/bogus-not-registered')
+    assert rv.status_code == 404